finished work

+++ backend/src/main/java/ec/edu/ups/clinic/backend/services/SpecialtyServiceREST.java
@@ -1,6 +1,8 @@
 package ec.edu.ups.clinic.backend.services;
 
+import ec.edu.ups.clinic.backend.dao.AppointmentDAO;
 import ec.edu.ups.clinic.backend.dao.SpecialtyDAO;
+import ec.edu.ups.clinic.backend.model.Appointment;
 import ec.edu.ups.clinic.backend.model.Specialty;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
@@ -14,7 +16,9 @@ public class SpecialtyServiceREST {
 
     @Inject
     private SpecialtyDAO specialtyDAO;
-
+    
+    @Inject
+    private AppointmentDAO appointmentDAO;
     @GET
     public List<Specialty> list() {
         return specialtyDAO.getAll();
@@ -41,6 +45,11 @@ public class SpecialtyServiceREST {
     @DELETE
     @Path("/{id}")
     public void delete(@PathParam("id") Long id) {
+        List<Appointment> citas = appointmentDAO.findBySpecialtyId(id);
+        if (!citas.isEmpty()) {
+            throw new WebApplicationException("No se puede eliminar: especialidad con citas agendadas", 400);
+        }
         specialtyDAO.delete(id);
     }
+
 }
